유저 mmr 타입 불일치 오류 수정

+++ api/unbound-service/src/main/java/com/badboys/unbound_service/model/ChatMemberDto.java
@@ -17,7 +17,7 @@ public class ChatMemberDto {
 
     private String profileImage;
 
-    private Long mmr;
+    private int mmr;
 
     private String lastReadMessageId;
 }
